breaks out queries
