+++ libs/data-fetch-utils/src/index.ts
@@ -1,4 +1,2 @@
-export * from './utils/fetch';
-export * from './utils/fetchErrors';
-export * from './utils/ipfs';
-export * from './utils/requests';
+export * from './utils';
+export * from './types';
