+++ libs/moloch-v3-data/src/index.ts
@@ -1,6 +1,7 @@
-import Haus from './Haus';
-
+export * from './daos';
+export * from './members';
+export * from './proposals';
+export * from './transactions';
 export * from './types';
 export * from './utils';
-
-export { Haus };
+export * from './vaults';
