--- libs/moloch-v3-data/test/ordering.spec.ts
@@ -1,12 +1,0 @@
-import { Haus } from '../src/index';
-
-describe('haus', () => {
-  let haus: Haus;
-
-  beforeAll(async () => {
-    haus = Haus.create();
-  });
-  it('should have a query class', () => {
-    expect(haus.query).toBeTruthy();
-  });
-});
